flaskserver/requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
ultralytics>=8.0.186
opencv-python>=4.8.1.78
pillow>=10.0.1
numpy>=1.26.0
torch>=2.1.0
